socorro

+++ server/src/main/java/dev/mixsource/port/output/repository/Characters.java
@@ -8,13 +8,12 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.Optional;
 
 public interface Characters extends JpaRepository<CharacterEntity, Long> {
     @Transactional
     @Modifying
     @Query("UPDATE CharacterEntity c SET c.x = :x, c.y = :y WHERE c.id = :id")
-    void updateXAndYById(@Param("id") Long id, @Param("x") Integer x, @Param("y") Integer y);
+    void updateXAndYById(@Param("id") Long id, @Param("x") Double x, @Param("y") Double y);
 
 
     List<CharacterEntity> findAllByUserId(final Long userId);
